requirements.txt
requests>=2.31.0
tenacity>=8.2.3
pyyaml>=6.0.1
pydantic>=2.0
jinja2>=3.1
aiohttp>=3.9.0
